+++ src/main/java/io/electrum/moneytransfer/model/MoneyTransferQuoteRequest.java
@@ -24,6 +24,9 @@ public class MoneyTransferQuoteRequest {
    @JsonProperty("receiver")
    private Institution receiver = null;
 
+   @JsonProperty("settlementEntity")
+   private Institution settlementEntity = null;
+
    @JsonProperty("senderDetails")
    private PersonalDetails senderDetails;
 
@@ -78,6 +81,27 @@ public class MoneyTransferQuoteRequest {
 
    public void setReceiver(Institution receiver) {
       this.receiver = receiver;
+   }
+
+   public MoneyTransferQuoteRequest settlementEntity(Institution settlementEntity) {
+      this.settlementEntity = settlementEntity;
+      return this;
+   }
+
+   /**
+    * Data relating to the entity with whom the Merchant will settle the transaction..
+    *
+    * @return settlementEntity
+    **/
+   @JsonProperty("settlementEntity")
+   @ApiModelProperty(required = false, value = "Data relating to the entity with whom the Merchant will settle the transaction.")
+   @Valid
+   public Institution getSettlementEntity() {
+      return settlementEntity;
+   }
+
+   public void setSettlementEntity(Institution settlementEntity) {
+      this.settlementEntity = settlementEntity;
    }
 
    public MoneyTransferQuoteRequest senderDetails(PersonalDetails senderDetails) {
@@ -178,6 +202,7 @@ public class MoneyTransferQuoteRequest {
       MoneyTransferQuoteRequest moneyTransferFeeQuote = (MoneyTransferQuoteRequest) o;
       return Objects.equals(this.originator, moneyTransferFeeQuote.originator)
             && Objects.equals(this.receiver, moneyTransferFeeQuote.receiver)
+            && Objects.equals(this.settlementEntity, moneyTransferFeeQuote.settlementEntity)
             && Objects.equals(this.senderDetails, moneyTransferFeeQuote.senderDetails)
             && Objects.equals(this.recipientDetails, moneyTransferFeeQuote.recipientDetails)
             && Objects.equals(this.amount, moneyTransferFeeQuote.amount)
@@ -186,7 +211,8 @@ public class MoneyTransferQuoteRequest {
 
    @Override
    public int hashCode() {
-      return Objects.hash(originator, receiver, senderDetails, recipientDetails, amount, amountIncludesFees);
+      return Objects
+            .hash(originator, receiver, settlementEntity, senderDetails, recipientDetails, amount, amountIncludesFees);
    }
 
    @Override
@@ -195,6 +221,7 @@ public class MoneyTransferQuoteRequest {
       sb.append("class MoneyTransferQuoteRequest {\n");
       sb.append("    originator: ").append(Utils.toIndentedString(originator)).append("\n");
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
+      sb.append("    settlementEntity: ").append(Utils.toIndentedString(settlementEntity)).append("\n");
       sb.append("    senderDetails: ").append(Utils.toIndentedString(senderDetails)).append("\n");
       sb.append("    recipientDetails: ").append(Utils.toIndentedString(recipientDetails)).append("\n");
       sb.append("    amount: ").append(Utils.toIndentedString(amount)).append("\n");